pyproject.toml
[project]
name = "easypaper"
version = "0.1.0"
description = "AI-powered academic paper generation SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.5",
    "openai>=2.7.2",
    "langgraph>=0.0.26",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "pymupdf>=1.26.7",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.121.1",
    "uvicorn[standard]>=0.30.0,<0.31",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ipython>=9.7.0",
]
vlm = [
    "anthropic>=0.18.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["easypaper", "easypaper.*"]

[tool.uv.workspace]
members = ["user_case"]
